modify: 1. week1 solution이랑 test 순서 바뀐거 fixed 2. update week2 solution and test codes
