+++ GaHee/Week1/DevFunction/solution.test.js
@@ -1,13 +1,10 @@
 const solution = require('./solution.js');
 
-describe('가장 큰 수 Test', () => {
-  it('입력값이 [6, 10, 2] 일 때 기대값은 "6210" 이어야 합니다.', () => {
-    expect(solution([6, 10, 2])).toStrictEqual('6210');
+describe('기능 개발 test', () => {
+  it('입력값이 [93, 30, 55], [1, 30, 5] 일 때 기대값은 [2, 1] 이어야 합니다.', () => {
+    expect(solution([93, 30, 55], [1, 30, 5])).toStrictEqual([2, 1]);
   });
-  it('입력값이 [3, 30, 34, 5, 9] 일 때 기대값은 "9534330" 이어야 합니다.', () => {
-    expect(solution([3, 30, 34, 5, 9])).toStrictEqual('9534330');
-  });
-  it('입력값이 [0, 0, 0, 0] 일 때 기대값은 "0" 이어야 합니다.', () => {
-    expect(solution([0, 0, 0, 0])).toStrictEqual('0');
+  it('입력값이 [95, 90, 99, 99, 80, 99], [1, 1, 1, 1, 1, 1] 일 때 기대값은 [1, 3, 2] 이어야 합니다.', () => {
+    expect(solution([95, 90, 99, 99, 80, 99], [1, 1, 1, 1, 1, 1])).toStrictEqual([1, 3, 2]);
   });
 });
